Fix unreliability in verify tests
In https://bugs.debian.org/922259, Santiago Vila reported that the
verify tests occasionally fail like this:

  Traceback (most recent call last):
    File "tests/unit/test_verify.py", line 30, in test_verify32
      v32x[libnacl.randombytes_random() & 31] += 1
  ValueError: byte must be in range(0, 256)

And indeed you can reproduce this by just running those tests in a loop
(e.g. `while python -m unittest discover --start-directory tests -v -p
test_verify.py; do :; done`).

This happens in the corner case where the previous value of the byte
that the test tries to increment happens to be 255; in that case it will
effectively try to assign 256 to a bytearray element and fail.  We can
fix this by explicitly folding the value back into the range of a byte.